Abstract adaptor, renamed methods, minimum level
Created an abstract adaptor class to be extended. It provides functions
such as set/get date format, set/get last log line, and set minimum
log level for the adaptor. It also has an isHandling() method to
check if the adaptor is set to handle a specific log level.

A few methods have been renamed for consistancy and to better
express what they do.

A minimum log level system has been added. An adaptor can be set to
only handle levels above a specified level. For example, setting it
to warning will not handle logs for notice, info, or debug.