Named logger, adaptors; remove adaptor; ChromeLogger levels to ''
Added get/set name methods to the logger and adaptors. This is used
with the remove adaptor as an easy way to remove adaptors by name
if needed.

Changed DEBUG and NOTICE levels to '' in ChromeLogger which maps to
a 'log' type. Any empty string helps save bandwidth. Small, but some.